Add support for 128 bit decimal add without downcasting first
