+++ sql-plugin/src/main/340+/scala/org/apache/spark/sql/rapids/arithmetic.scala
@@ -19,15 +19,13 @@ package org.apache.spark.sql.rapids
 import scala.math.{max, min}
 
 import ai.rapids.cudf._
-import ai.rapids.cudf.ast.BinaryOperator
 import com.nvidia.spark.rapids._
-import com.nvidia.spark.rapids.shims.GpuTypeShims
 
 import org.apache.spark.sql.catalyst.analysis.{DecimalPrecision, TypeCheckResult}
 import org.apache.spark.sql.catalyst.expressions.{Expression, NullIntolerant}
 import org.apache.spark.sql.internal.SQLConf
-import org.apache.spark.sql.rapids.shims.RapidsErrorUtils
 import org.apache.spark.sql.types._
+import org.apache.spark.sql.vectorized.ColumnarBatch
 
 abstract class CudfBinaryArithmetic extends CudfBinaryOperator with NullIntolerant {
 
@@ -56,48 +54,79 @@ abstract class CudfBinaryArithmetic extends CudfBinaryOperator with NullIntolera
   }
 }
 
+trait GpuAddSub extends CudfBinaryArithmetic {
+
+  override def outputTypeOverride: DType = GpuColumnVector.getNonNestedRapidsType(dataType)
+  val failOnError: Boolean
+
+  override def columnarEval(batch: ColumnarBatch): Any = {
+    if (dataType.isInstanceOf[DecimalType]) {
+      (left.dataType, right.dataType) match {
+        case (DecimalType.Fixed(p1, s1), DecimalType.Fixed(p2, s2)) =>
+          val resultType = resultDecimalType(p1, s1, p2, s2)
+          if (resultType.precision < 38) {
+            super.columnarEval(batch)
+          } else {
+            // call the kernel to add 128-bit numbers
+            prepareInputAndExecute(batch, resultType)
+          }
+        case _ => throw new IllegalStateException("Both operands should be Decimal")
+      }
+    } else {
+      super.columnarEval(batch)
+    }
+  }
+
+  protected def prepareInputAndExecute(
+      batch: ColumnarBatch,
+      resultType: DecimalType): Any = {
+    val castLhs = withResource(GpuExpressionsUtils.columnarEvalToColumn(left, batch)) { lhs =>
+      lhs.getBase.castTo(DType.create(DType.DTypeEnum.DECIMAL128, lhs.getBase.getType.getScale))
+    }
+    val retTab = withResource(castLhs) { castLhs =>
+      val castRhs = withResource(GpuExpressionsUtils.columnarEvalToColumn(right, batch)) { rhs =>
+        rhs.getBase.castTo(DType.create(DType.DTypeEnum.DECIMAL128, rhs.getBase.getType.getScale))
+      }
+      withResource(castRhs) { castRhs =>
+        do128BitOperation(castLhs, castRhs, -resultType.scale)
+      }
+    }
+    val retCol = withResource(retTab) { retTab =>
+      val overflowed = retTab.getColumn(0)
+      val sum = retTab.getColumn(1)
+      if (failOnError) {
+        withResource(overflowed.any()) { anyOverflow =>
+          if (anyOverflow.isValid && anyOverflow.getBoolean) {
+            throw new ArithmeticException(GpuCast.INVALID_INPUT_MESSAGE)
+          }
+        }
+        sum.incRefCount()
+      } else {
+        withResource(GpuScalar.from(null, dataType)) { nullVal =>
+          overflowed.ifElse(nullVal, sum)
+        }
+      }
+    }
+    GpuColumnVector.from(retCol, dataType)
+  }
+
+  protected def do128BitOperation(
+      castLhs: ColumnView,
+      castRhs: ColumnView,
+      outputScale: Int): Table
+}
+
 case class GpuAdd(
     left: Expression,
     right: Expression,
-    failOnError: Boolean) extends CudfBinaryArithmetic {
-  override def inputType: AbstractDataType = TypeCollection.NumericAndInterval
+    override val failOnError: Boolean)
+    extends GpuAddParent(left, right, failOnError) with GpuAddSub {
 
-  override def symbol: String = "+"
-
-  override def binaryOp: BinaryOp = BinaryOp.ADD
-
-  override def astOperator: Option[BinaryOperator] = Some(ast.BinaryOperator.ADD)
-
-  override def doColumnar(lhs: BinaryOperable, rhs: BinaryOperable): ColumnVector = {
-    val ret = super.doColumnar(lhs, rhs)
-    withResource(ret) { ret =>
-      // No shims are needed, because it actually supports ANSI mode from Spark v3.0.1.
-      if (failOnError && GpuAnsi.needBasicOpOverflowCheck(dataType) ||
-          GpuTypeShims.isSupportedDayTimeType(dataType) ||
-          GpuTypeShims.isSupportedYearMonthType(dataType)) {
-        // For day time interval, Spark throws an exception when overflow,
-        // regardless of whether `SQLConf.get.ansiEnabled` is true or false
-        AddOverflowChecks.basicOpOverflowCheck(lhs, rhs, ret)
-      }
-
-      if (dataType.isInstanceOf[DecimalType]) {
-        val sparkDecimal = dataType.asInstanceOf[DecimalType]
-        val cudfDecimal = DecimalUtil.createCudfDecimal(sparkDecimal)
-        withResource {
-          if (!cudfDecimal.equals(ret.getType)) {
-            withResource(ret.round(sparkDecimal.scale)) { rounded =>
-              rounded.castTo(cudfDecimal)
-            }
-          } else {
-            ret.incRefCount()
-          }
-        } { ret =>
-          AddOverflowChecks.decimalOpOverflowCheck(lhs, rhs, ret, failOnError)
-        }
-      } else {
-        ret.incRefCount()
-      }
-    }
+  def do128BitOperation(
+      castLhs: ColumnView,
+      castRhs: ColumnView,
+      outputScale: Int): Table = {
+    com.nvidia.spark.rapids.jni.DecimalUtils.add128(castLhs, castRhs, outputScale)
   }
 
   // scalastyle:off
@@ -121,101 +150,14 @@ case class GpuAdd(
 case class GpuSubtract(
     left: Expression,
     right: Expression,
-    failOnError: Boolean) extends CudfBinaryArithmetic {
-  override def inputType: AbstractDataType = TypeCollection.NumericAndInterval
+    override val failOnError: Boolean)
+    extends GpuSubtractParent(left, right, failOnError) with GpuAddSub {
 
-  override def symbol: String = "-"
-
-  override def binaryOp: BinaryOp = BinaryOp.SUB
-  override def astOperator: Option[BinaryOperator] = Some(ast.BinaryOperator.SUB)
-
-  private[this] def basicOpOverflowCheck(
-      lhs: BinaryOperable,
-      rhs: BinaryOperable,
-      ret: ColumnVector): Unit = {
-    // Check overflow. It is true if the arguments have different signs and
-    // the sign of the result is different from the sign of x.
-    // Which is equal to "((x ^ y) & (x ^ r)) < 0" in the form of arithmetic.
-
-    val signCV = withResource(lhs.bitXor(rhs)) { xyXor =>
-      withResource(lhs.bitXor(ret)) { xrXor =>
-        xyXor.bitAnd(xrXor)
-      }
-    }
-    val signDiffCV = withResource(signCV) { sign =>
-      withResource(Scalar.fromInt(0)) { zero =>
-        sign.lessThan(zero)
-      }
-    }
-    withResource(signDiffCV) { signDiff =>
-      withResource(signDiff.any()) { any =>
-        if (any.isValid && any.getBoolean) {
-          throw RapidsErrorUtils.arithmeticOverflowError(
-            "One or more rows overflow for Subtract operation."
-          )
-        }
-      }
-    }
-  }
-
-  private[this] def decimalOpOverflowCheck(
-      lhs: BinaryOperable,
-      rhs: BinaryOperable,
-      ret: ColumnVector): ColumnVector = {
-    // We need a special overflow check for decimal because CUDF does not support INT128 so we
-    // cannot reuse the same code for the other types.
-    // Overflow happens if the arguments have different signs and the sign of the result is
-    // different from the sign of subtractend (RHS).
-    val numRows = ret.getRowCount.toInt
-    val zero = BigDecimal(0).bigDecimal
-    val overflow = withResource(DecimalUtils.lessThan(rhs, zero, numRows)) { rhsLz =>
-      val argsSignDifferent = withResource(DecimalUtils.lessThan(lhs, zero, numRows)) { lhsLz =>
-        lhsLz.notEqualTo(rhsLz)
-      }
-      withResource(argsSignDifferent) { argsSignDifferent =>
-        val resultAndSubtrahendSameSign =
-          withResource(DecimalUtils.lessThan(ret, zero)) { resultLz =>
-            rhsLz.equalTo(resultLz)
-          }
-        withResource(resultAndSubtrahendSameSign) { resultAndSubtrahendSameSign =>
-          resultAndSubtrahendSameSign.and(argsSignDifferent)
-        }
-      }
-    }
-    withResource(overflow) { overflow =>
-      if (failOnError) {
-        withResource(overflow.any()) { any =>
-          if (any.isValid && any.getBoolean) {
-            throw new ArithmeticException("One or more rows overflow for Subtract operation.")
-          }
-        }
-        ret.incRefCount()
-      } else {
-        withResource(GpuScalar.from(null, dataType)) { nullVal =>
-          overflow.ifElse(nullVal, ret)
-        }
-      }
-    }
-  }
-
-  override def doColumnar(lhs: BinaryOperable, rhs: BinaryOperable): ColumnVector = {
-    val ret = super.doColumnar(lhs, rhs)
-    withResource(ret) { ret =>
-      // No shims are needed, because it actually supports ANSI mode from Spark v3.0.1.
-      if (failOnError && GpuAnsi.needBasicOpOverflowCheck(dataType) ||
-          GpuTypeShims.isSupportedDayTimeType(dataType) ||
-          GpuTypeShims.isSupportedYearMonthType(dataType)) {
-        // For day time interval, Spark throws an exception when overflow,
-        // regardless of whether `SQLConf.get.ansiEnabled` is true or false
-        basicOpOverflowCheck(lhs, rhs, ret)
-      }
-
-      if (dataType.isInstanceOf[DecimalType]) {
-        decimalOpOverflowCheck(lhs, rhs, ret)
-      } else {
-        ret.incRefCount()
-      }
-    }
+  def do128BitOperation(
+      castLhs: ColumnView,
+      castRhs: ColumnView,
+      outputScale: Int): Table = {
+    com.nvidia.spark.rapids.jni.DecimalUtils.subtract128(castLhs, castRhs, outputScale)
   }
 
   override def resultDecimalType(p1: Int, s1: Int, p2: Int, s2: Int): DecimalType = {
@@ -312,7 +254,7 @@ case class GpuIntegralDivide(
 case class GpuRemainder(
     left: Expression,
     right: Expression)
-    extends GpuRemainderParent(left, right) with DecimalWithPromote {
+    extends GpuRemainderParent(left, right) {
 
   // scalastyle:off
   // The formula follows Hive which is based on the SQL standard and MS SQL:
@@ -334,7 +276,7 @@ case class GpuRemainder(
 
 case class GpuPmod(
     left: Expression,
-    right: Expression) extends GpuPmodParent(left, right) with DecimalWithPromote {
+    right: Expression) extends GpuPmodParent(left, right) {
   // This follows Remainder rule
   override def resultDecimalType(p1: Int, s1: Int, p2: Int, s2: Int): DecimalType = {
     val resultScale = max(s1, s2)
